backend/requirements.txt
fastapi
uvicorn
sqlalchemy
geoalchemy2
psycopg2-binary

pymilvus>=2.5.0
protobuf==5.29.1






torch
transformers
pillow
python-multipart
requests
numpy
